A bunch of work on trying to make the Z-Machine multiplayer.
Not quite there; we need to deal with get_prop_addr and such not having a
valid spot in dynamic memory for multiplayer objects, which crashes the
Z-machine pretty quickly, and we probably have to mess with more global vars
than just G00.

But this is progress!

You can have two people walk in separate directions in parallel, so we're
doing okay as a start here.