Move multiplayer object property data to fake Z-Machine addresses.
They need to fit in a uint16, but there isn't any memory obviously available,
so put them at the end of the address space, in what _should_ be illegal for
the game to access (code sections). When we see these illegal accesses, fix
them up to our external memory instead of the actual Z-Machine memory.

This gets it to the point where two people can walk around and interact with
the world independently!