Make sure the PLAYER global is set to the multiplayer object number.
Otherwise things that compare this number will fail, since we can catch
accesses on specific objects but not integer comparisons.

Notably: when you go to fight the troll, it will look for the sword in
PLAYER's inventory, and not find it, because PLAYER would be the
original "cretin" object.

(and probably many other things like this.)

Fixes #7.